// web/expo-money/src/types/investmenstDTO.tsx
export interface InvestmentsDTO {
    investmentsValue: number;
    totalExpectedValue: number;
    profitValue: number;
    delayedExpectedValue: number;
    formatedInvestmentsValue: string;
    formatedTotalExpectedValue: string;
    formatedProfitValue: string;
    formatedDelayedExpectedValue: string;
}